renaming of challenge cards
