+++ client/src/components/mychallenges/mychallenges.js
@@ -2,7 +2,7 @@ import React from 'react';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import Typography from '@material-ui/core/Typography';
 import Container from '@material-ui/core/Container';
-import ChallengeCard from "../challengecard/challengecard";
+import UserChallenge from "../userchallenge/userchallenge";
 import Modal from "../modal/modal";
 import "./mychallenges.css";
 import Paper from '@material-ui/core/Paper';
@@ -16,7 +16,7 @@ function MyChallenges() {
                     <Typography component="div"
                         style={{ backgroundColor: '#cfe8fc', height: '100vh', padding: 30, paddingTop: 20, fontSize: 30 }}>
                     <Modal />
-                        <ChallengeCard />
+                        <UserChallenge />
                     </Typography>
                 </Paper>
             </Container>
